Fix type-casting

+++ milestone3/tests/test_3.java
@@ -3,7 +3,7 @@ class test_25
     public static void main(String args[])
     {  
 		int x = 20;
-		int y = 18;
+		float y = 18;
 		if (x > y) {
 			println("x is greater than y");
 		}
